fix(secureZoneRepo): deleteAllKeys 임시 삭제
EncryptedSharedPreferences)가 쓰는 AndroidX Security의 마스터키(alias: androidx_security_master_key)가 삭제되는 버그 수정

+++ lib/repository/old_secure_storage_cleaner_for_ios.dart
@@ -1,13 +1,18 @@
+import 'dart:io';
+
+import 'package:coconut_vault/repository/secure_storage_repository.dart';
 import 'package:coconut_vault/utils/logger.dart' show Logger;
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
 
-class OldSecureStorageCleaner {
+class OldSecureStorageCleanerForIos {
   // 이전 설정
   static const FlutterSecureStorage _oldStorage = FlutterSecureStorage(
-    aOptions: AndroidOptions(encryptedSharedPreferences: false),
+    aOptions: SecureStorageRepository.androidOptions,
   );
 
   static Future<void> cleanAll() async {
+    if (!Platform.isIOS) return;
+
     try {
       await _oldStorage.deleteAll();
       final Map<String, String> oldData = await _oldStorage.readAll();
